+++ api/src/database/resources/mod.rs
@@ -1,0 +1,2 @@
+pub mod flavor;
+pub mod flavor_group;