+++ api/src/database/user/mod.rs
@@ -1,0 +1,3 @@
+pub mod project;
+#[allow(clippy::module_inception)]
+pub mod user;