+++ api/src/routes/budgeting/project_budget/mod.rs
@@ -1,13 +1,14 @@
 use actix_web::web::{
     delete,
-    // get, patch, post,
+    // get, patch,
+    post,
     scope,
 };
 use actix_web::Scope;
 use serde::Deserialize;
 
-// mod create;
-// use create::project_budget_create;
+mod create;
+use create::project_budget_create;
 // mod list;
 // use list::project_budget_list;
 // mod get;
@@ -19,7 +20,7 @@ use delete::project_budget_delete;
 
 pub fn project_budgets_scope() -> Scope {
     scope("/projectbudgets")
-        // .route("/", post().to(project_budget_create))
+        .route("/", post().to(project_budget_create))
         // .route("", get().to(project_budget_list))
         // .route("/{project_budget_id}", get().to(project_budget_get))
         // TODO: what about PUT?