+++ api/src/routes/mod.rs
@@ -16,15 +16,6 @@ pub use quota::*;
 pub use resources::*;
 pub use user::*;
 
-// TODO: create endpoints for
-// - accounting::server_state
-// - resources::flavor_group
-// - resources::flavor
-// - quota::flavor_quota
-// - pricing::flavor_price
-// - budgeting::project_budget
-// - budgeting::user_budget
-
 // TODO: modify endpoints for
 // - accounting::server_state
 // - resources::flavor_group