+++ api/src/routes/quota/flavor_quota/mod.rs
@@ -1,13 +1,14 @@
 use actix_web::web::{
     delete,
-    // get, patch, post,
+    // get, patch,
+    post,
     scope,
 };
 use actix_web::Scope;
 use serde::Deserialize;
 
-// mod create;
-// use create::flavor_quota_create;
+mod create;
+use create::flavor_quota_create;
 // mod list;
 // use list::flavor_quota_list;
 // mod get;
@@ -19,7 +20,7 @@ use delete::flavor_quota_delete;
 
 pub fn flavor_quotas_scope() -> Scope {
     scope("/flavorquotas")
-        // .route("/", post().to(flavor_quota_create))
+        .route("/", post().to(flavor_quota_create))
         // .route("", get().to(flavor_quota_list))
         // .route("/{flavor_quota_id}", get().to(flavor_quota_get))
         // TODO: what about PUT?